+++ webapp/src/components/user_attribute/user_attribute.jsx
@@ -4,18 +4,18 @@ import PropTypes from 'prop-types';
 import {describeTenure} from '../../dateutil';
 import {id as pluginId} from '../../manifest';
 
-const {messageHtmlToComponent} = window.PostUtils;
+const {messageHtmlToComponent, formatText} = window.PostUtils;
 
 export default class UserAttribute extends React.PureComponent {
     static propTypes = {
-        email: PropTypes.string,
+        username: PropTypes.string,
         pingboardInfo: PropTypes.object,
         fetchAndStorePingboardInfo: PropTypes.func.isRequired,
     }
 
     constructor(props) {
         super(props);
-        props.fetchAndStorePingboardInfo(props.email);
+        props.fetchAndStorePingboardInfo(props.username);
     }
 
     render() {
@@ -27,10 +27,15 @@ export default class UserAttribute extends React.PureComponent {
         const startDate = new Date(pingboardInfo.start_year, pingboardInfo.start_month - 1, pingboardInfo.start_day);
         const tenure = describeTenure(startDate, localDate);
         const description = pingboardInfo.job_title + (pingboardInfo.department ? ` (${pingboardInfo.department})` : '');
+        const manager = pingboardInfo.manager ? `@${pingboardInfo.manager}` : '(unknown manager)';
+
         return (
             <div>
                 <div key={`${pluginId}_job_title`}>
                     {messageHtmlToComponent(`👤 ${description}`)}
+                </div>
+                <div key={`${pluginId}_manager`}>
+                    {messageHtmlToComponent(formatText(`⬆️ ${manager}`, {atMentions: true, emoticons: false}))}
                 </div>
                 <div key={`${pluginId}_start_date`}>
                     {messageHtmlToComponent(`🗓 ${tenure}`)}