+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## 0.0.1 - 2018-08-16
-### Added
+## 0.0.1 - 2020-05-20
+
 - Initial release
+
+## 0.0.2 - 2020-12-19
+
+- Normalise emails when matching mattermost and pingboard users
+- Add manager to popover